IndianCovidFighters
Our Country is facing the deadly 2nd wave of covid-19 and we would like to contribute something for our country so, We are forwarding you here with an appeal for donation by IndianCovidfighters, to all the people reaching out to this article for supporting and generously contributing as donation for controlling the pandemic and allied relief work for the needy section in the society. we at IndianCovidfighters are raising funds to help the people who are suffering for plasma, oxygen, food for covid patients, medicines, important injections, beds availability, online doctor appointments etc. We humbly request you to generously donate for this Nobel cause.